% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{add.cell.type}
\alias{add.cell.type}
\title{Add new cell type}
\usage{
add.cell.type(
  type_name,
  valence,
  temporal_modulation_bias = 0.001,
  temporal_modulation_timeconstant = 1,
  temporal_modulation_amplitude = 0.005,
  transmission_velocity = 30000,
  v_bound = 85,
  dHdv_bound = 1.05e-06,
  I_spike = 1e-06,
  coupling_scaling_factor = 1e-07,
  spike_potential = 35,
  resting_potential = -70,
  threshold = -55
)
}
\arguments{
\item{type_name}{Character string giving name of the cell type, e.g. "excitatory", "inhibitory", "PV", "SST", etc.}

\item{valence}{Valence of each neuron type, +1 for excitatory, -1 for inhibitory}

\item{temporal_modulation_bias}{Temporal modulation time (in ms) bias for each neuron type. Default value is 1e-3.}

\item{temporal_modulation_timeconstant}{Temporal modulation time (in ms) step for each neuron type. Default value is 1e0.}

\item{temporal_modulation_amplitude}{Temporal modulation time (in ms) cutoff for each neuron type. Default value is 5e-3.}

\item{transmission_velocity}{Transmission velocity (in microns/ms) for each neuron type. Default value is 30e3.}

\item{v_bound}{Potential bound, such that -v_bound <= v_traces <= v_bound, in unit_potential (mV), for each neuron in the network, based on its type. Default value is 85.0.}

\item{dHdv_bound}{Bound on derivative of metabolic energy wrt potential, such that dHdv_bound > abs(dHdv), in mA, for each neuron in the network, based on its type. Default value is 1.05e-6.}

\item{I_spike}{Spike current, in mA. Default value is 1e-6 (i.e., 1 nA).}

\item{coupling_scaling_factor}{Controls how energy used in synaptic transmission compares to that used in spiking. Default value is 1e-7, meaning that synaptic transmission uses 0.00001 percent of the energy used in spiking.}

\item{spike_potential}{Magnitude of each spike, in mV. Default value is 35.0.}

\item{resting_potential}{Resting potential, in mV. Default value is -70.0.}

\item{threshold}{Spike threshold, in mV. Default value is -55.0.}
}
\value{
Nothing.
}
\description{
This function adds a user-defined cell type to the current session. It's just a wrapper for the Rcpp-exported \code{add_cell_type} function. Technically, \code{cell_type} is a \code{struc} defined in the Rcpp backend of the neurons package. They are essentially labeled lists with the following entries: \code{type_name}, \code{valence}, \code{temporal_modulation_bias}, \code{temporal_modulation_timeconstant}, \code{temporal_modulation_amplitude}, \code{transmission_velocity}, \code{v_bound}, \code{dHdv_bound}, \code{I_spike}, \code{coupling_scaling_factor}, \code{spike_potential}, \code{resting_potential}, and \code{threshold}. Each session stores cell types in the Rcpp backend in an \code{unordered_map} with \code{string} labels. All parameters come with biologically realistic (and mathematically workable) default values, except for \code{type_name} and \code{valence}.
}
