% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{analyze.autocorr}
\alias{analyze.autocorr}
\title{Function to analyze autocorrelation estimates with bootstrapping}
\usage{
analyze.autocorr(ests, covariate, n_bs = 10000)
}
\arguments{
\item{ests}{Output from \code{estimate.autocorr.params}, or a list of such outputs.}

\item{covariate}{Character string or vector of character strings specifying the covariate(s) to analyze (e.g., "hemi").}

\item{n_bs}{Number of bootstrap resamples to perform (default: 1e4).}
}
\value{
A list with two elements: 
 \describe{
   \item{resamples}{data frame with bootstrap resamples of the mean tau for each combination of covariate levels}
   \item{distribution_plot}{ggplot object showing the distribution of mean tau for each combination of covariate levels}
 }
}
\description{
This function takes the output from \code{estimate.autocorr.params} and performs a bootstrap analysis to estimate the distribution of the mean time constant (tau) for different levels of specified covariates.
}
