% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{fetch.cell.type.params}
\alias{fetch.cell.type.params}
\title{Fetch cell type parameters}
\usage{
fetch.cell.type.params(type_name)
}
\value{
List of parameters for the named cell type.
}
\description{
This function returns the parameters for a named cell type in a list. It's just a wrapper for the Rcpp-exported \code{fetch_cell_type_params} function.
}
