% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{fit.edf.autocorr}
\alias{fit.edf.autocorr}
\title{Fit exponential decay function to autocorrelation of single neuron object}
\usage{
fit.edf.autocorr(nrn, A0 = 0.001, tau0 = 1, ctol = 1e-08, max_evals = 500)
}
\arguments{
\item{nrn}{Neuron object for which to fit autocorrelation.}

\item{A0}{Initial guess for amplitude parameter of exponential decay function (default: 0.001).}

\item{tau0}{Initial guess for time constant parameter of exponential decay function (default: 1.0).}

\item{ctol}{Convergence tolerance for fitting exponential decay function (default: 1e-8).}

\item{max_evals}{Maximum number of evaluations for fitting exponential decay function (default: 500).}
}
\value{
None. Modifies neuron in place.
}
\description{
This function fits an exponential decay function to the autocorrelation of a single neuron object using specified parameters. It's a wrapper around the \code{fit_autocorrelation} method of the neuron class.
}
