% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{new.motif}
\alias{new.motif}
\title{Initialize motif}
\usage{
new.motif(motif_name = "not_provided")
}
\arguments{
\item{motif_name}{Character string giving name of the motif (default: "not_provided").}
}
\value{
A new motif object.
}
\description{
This function initializes a new motif object with specified parameters. Motifs are used for building networks of interconnected neurons. They are recipes for building internode projections within a neural network. They are "columnar", in the sense that they are repeated across cortical columns.
}
