% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{plot.network.traces}
\alias{plot.network.traces}
\title{Plot spike traces for network from SGT simulation}
\usage{
plot.network.traces(network, return_plot)
}
\arguments{
\item{network}{Network object with SGT simulation traces to plot.}

\item{return_plot}{Logical indicating whether to return the ggplot object (TRUE) or print it (FALSE) (default: FALSE).}
}
\value{
A ggplot object showing spike traces for all neurons in the network over time.
}
\description{
This function plots spike traces for a network object from a Spatial Growth-Transform (SGT) simulation.
}
