% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{plot.network}
\alias{plot.network}
\title{Plot network as directed graph}
\usage{
plot.network(network, title = NULL, plot_motif = "local connections", edge_color = "pre_type", cell_color = "layer", cell_size_factor = 5.0, arrow_size_factor = 0.5, return_plot = FALSE)
}
\arguments{
\item{network}{Network object to plot.}

\item{title}{Title for the plot (default: "Cortical Patch" or network name (if provided), plus plot motif name(s)).}

\item{plot_motif}{Character string specifying which motif to plot; options include "local" for local connections within each node or the name of a long-range projection motif (default: "local connections").}

\item{edge_color}{Character string specifying how to color the edges; options include "pre_type" to color by presynaptic neuron type, "post_type" to color by postsynaptic neuron type, or "motif" to color by motif type (default: "pre_type").}

\item{cell_color}{Character string specifying how to color the nodes; options include "layer" to color by layer index or "type" to color by neuron type (default: "layer").}

\item{cell_size_factor}{Numeric value controlling how cell size in the plot scales to the number of cells.}

\item{return_plot}{Logical indicating whether to return the ggplot object (TRUE) or print the plot directly (FALSE) (default: FALSE).}

\item{arrow_size}{Numeric value indicating the size of the arrows representing edges (default: 0.05).}
}
\value{
Either prints the plot directly or returns the ggplot object, depending on the value of return_plot.
}
\description{
This function plots a network object as a directed graph using ggplot2. Nodes represent neurons, and directed edges represent connections between them. The plot can be customized by selecting which motif to display and how to color the edges.
}
