% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{plot.raster}
\alias{plot.raster}
\title{Function to plot spike raster of neuron}
\usage{
plot.raster(nrn, plot_title = "Spike raster", zero_as_onset = TRUE)
}
\arguments{
\item{nrn}{Neuron object for which to plot spike raster.}

\item{plot_title}{Title for the plot (default: "Spike raster").}

\item{zero_as_onset}{Logical indicating whether to plot a vertical line at time zero to indicate stimulus onset (default: TRUE).}
}
\value{
A ggplot object showing the spike raster.
}
\description{
Wrapper function to fetch and plot spike raster from a neuron object. Assumes the spike raster has already been loaded.
}
