% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{plot.network}
\alias{plot.network}
\title{Plot network as directed graph}
\usage{
\method{plot}{network}(
  network,
  title = "Cortical Patch",
  plot_motif = "local",
  edge_color = "pre_type",
  return_plot = FALSE
)
}
\arguments{
\item{network}{Network object to plot.}

\item{title}{Title for the plot (default: "Cortical Patch").}

\item{plot_motif}{Character string specifying which motif to plot; options include "local" for local connections within each node or the name of a long-range projection motif (default: "local").}

\item{edge_color}{Character string specifying how to color the edges; options include "pre_type" to color by presynaptic neuron type, "post_type" to color by postsynaptic neuron type, or "motif" to color by motif type (default: "pre_type").}
}
\description{
This function plots a network object as a directed graph using ggplot2. Nodes represent neurons, and directed edges represent connections between them. The plot can be customized by selecting which motif to display and how to color the edges.
}
