% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess.kilo4}
\alias{preprocess.kilo4}
\title{Preprocess kilosort4 data}
\usage{
preprocess.kilo4(
  trial_time_start = -100,
  trial_time_end = 2020,
  recording.folder = "data",
  meta_data = NULL,
  pure_trials_only = TRUE,
  good_cells_only = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{trial_time_start}{Time to begin trial (ms), relative to stimulus onset}

\item{trial_time_end}{Time to end trial (ms), relative to stimulus onset}

\item{recording.folder}{List of paths to the kilosort4 output folders}

\item{meta_data}{Data frame with metadata for each recording (e.g., genotype, hemisphere), one recording per row; row names should match recording names and all columns should be covariates for later analysis}

\item{pure_trials_only}{Keep only trials with no overlap?}

\item{good_cells_only}{Keep only cells marked as "good" in the cluster_group.tsv file?}

\item{verbose}{Print report on imported data?}
}
\value{
A list with three elements: 
 \describe{
   \item{spikes}{data frame with one row per spike}
   \item{timeXtrial}{list of matrices with rows as sample times and columns as trials, each element a zero if no spike at that time in that trial, and a one if a spike}
   \item{cluster.key}{data frame with one row per neuron}
 }
}
\description{
Function to preprocess kilosort4 data for use with neurons package functions.
}
