% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{print.known.celltypes}
\alias{print.known.celltypes}
\title{Print known cell types}
\usage{
print.known.celltypes()
}
\value{
Nothing.
}
\description{
This function prints names and all parameters for all cell types recognized in the current session. It's just a wrapper for the Rcpp-exported \code{print_known_celltypes} function.
}
