% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{run.GTsim}
\alias{run.GTsim}
\title{Run Growth-Transform network simulation}
\usage{
run.GTsim(network, stimulus_current_matrix, dt = 0.001)
}
\arguments{
\item{network}{Network object on which to run the simulation.}

\item{stimulus_current_matrix}{Matrix of input currents, with rows representing neurons and columns representing sample times.}

\item{dt}{Time step length in the unit_time of the network (default: 1e-3, or 1-micron time steps).}

\item{v_ceiling}{Potential ceiling in the unit_potential of the network (default: 1.0).}

\item{I_ceiling}{Current ceiling in the unit_current of the network (default: 1.0).}

\item{I_spike}{Spike current in the unit_current of the network (default: 5.0).}

\item{spike_potential}{Magnitude of each spike (default: 1.0).}

\item{resting_potential}{Resting potential in the unit_potential of the network (default: -70.0).}

\item{threshold}{Spike threshold in the unit_potential of the network (default: 0.0).}
}
\value{
A matrix containing the spike_traces of all neurons over time after the simulation (neurons as rows, sample times as columns).
}
\description{
This function runs a Growth-Transform network simulation on a given network object for a specified matrix of input currents over time and specified network electrical constants.
}
