% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{summarize.cluster.key}
\alias{summarize.cluster.key}
\title{Summarize cluster key by covariates}
\usage{
summarize.cluster.key(key, covariate_list)
}
\arguments{
\item{key}{The \code{cluster.key} data frame output of \code{preprocess.kilo4}}

\item{covariate_list}{List of covariates to summarize}
}
\value{
A data frame with one row per combination of covariates, and columns giving the number of cells, mean number of spikes, and mean number of trials for that combination
}
\description{
Function to summarize the \code{cluster.key} output of \code{preprocess.kilo4} by covariates.
}
