% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{test.sigma.assumption}
\alias{test.sigma.assumption}
\title{Helper function for testing assumption of autocorrelation processing}
\usage{
test.sigma.assumption(autocor_results)
}
\arguments{
\item{autocor_results}{Data frame of autocorrelation results from the \code{process.autocorr} function.}
}
\value{
Nothing; this function produces a base-R plot for visual inspection.
}
\description{
The approach we're using is valid if autocorr (red lines) is below mean firing rate (blue line). If so, then the equation we're using to convert spiking autocorr to guassian sigma should work.
}
