% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{load.projection.into.motif}
\alias{load.projection.into.motif}
\title{Load projection into motif}
\usage{
load.projection.into.motif(
  motif,
  presynaptic_type,
  presynaptic_layer,
  presynaptic_density,
  postsynaptic_type,
  postsynaptic_layer,
  postsynaptic_density,
  max_col_shift_up = 0,
  max_col_shift_down = 0,
  connection_strength = 1
)
}
\arguments{
\item{motif}{Motif object into which to load the projection.}

\item{presynaptic_type}{Character string giving type of presynaptic neuron, e.g. "excitatory", "inhibitory", etc.}

\item{presynaptic_layer}{Character string giving layer of presynaptic neuron, e.g. "L2/3", "L4", "L5", "L6", etc.}

\item{presynaptic_density}{Numeric giving density of presynaptic neuron type in presynaptic layer (e.g., ratio of neurons per node).}

\item{postsynaptic_type}{Character string giving type of postsynaptic neuron, e.g. "excitatory", "inhibitory", etc.}

\item{postsynaptic_layer}{Character string giving layer of postsynaptic neuron, e.g. "L2/3", "L4", "L5", "L6", etc.}

\item{postsynaptic_density}{Numeric giving density of postsynaptic neuron type in postsynaptic layer (e.g., ratio of neurons per node).}

\item{max_col_shift_up}{Maximum number of columns upwards (increasing columnar indexes) that the projection can reach (default: 0).}

\item{max_col_shift_down}{Maximum number of columns downwards (decreasing columnar indexes) that the projection can reach (default: 0).}

\item{connection_strength}{Numeric giving overall strength of the projection (default: 1.0).}
}
\value{
The updated motif object with the new projection loaded.
}
\description{
This function loads a projection schema into a motif object. Projections define internode connectivity within a network built using the motif.
}
