% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{plot.autocorrelation}
\alias{plot.autocorrelation}
\title{Function to plot autocorrelation of neuron}
\usage{
plot.autocorrelation(nrn, plot_title = "Est. autocorr", bias_term = 0, plot_time_cutoff = Inf)
}
\arguments{
\item{nrn}{Neuron object for which to plot autocorrelation.}

\item{plot_title}{Title for the plot (default: "Est. autocorr").}

\item{bias_term}{Bias term to plot as a horizontal line (default: NULL).}

\item{plot_time_cutoff}{Maximum lag (in bins) to display on the x-axis (default: Inf).}

\item{return_plot}{Logical indicating whether to return the ggplot object (TRUE) or print it (FALSE) (default: FALSE).}
}
\value{
A ggplot object showing the estimated and fitted autocorrelation.
}
\description{
Wrapper function to fetch and plot autocorrelation information from a neuron object. Assumes the autocorrelation has already been computed.
}
