% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.R
\name{run.GTsim}
\alias{run.GTsim}
\title{Run Growth-Transform network simulation}
\usage{
run.GTsim(
  network,
  stimulus_current_matrix,
  dt = 0.1,
  v_ceiling = 1,
  I_ceiling = 1,
  I_spike = 5,
  transimpedance = 1,
  threshold = 0
)
}
\arguments{
\item{network}{Network object on which to run the simulation.}

\item{stimulus_current_matrix}{Matrix of input currents, with rows representing neurons and columns representing sample times.}

\item{dt}{Time step length in the unit_time of the network (default: 0.1).}

\item{v_ceiling}{Potential ceiling in the unit_potential of the network (default: 1.0).}

\item{I_ceiling}{Current ceiling in the unit_current of the network (default: 1.0).}

\item{I_spike}{Spike current in the unit_current of the network (default: 5.0).}

\item{transimpedance}{Transimpedance value determining the magnitude of each spike (default: 1.0).}

\item{threshold}{Spike threshold in the unit_potential of the network (default: 0.0).}
}
\value{
A matrix containing the spike_traces of all neurons over time after the simulation (neurons as rows, sample times as columns).
}
\description{
This function runs a Growth-Transform network simulation on a given network object for a specified matrix of input currents over time and specified network electrical constants.
}
